More score -> points renaming.
I hope there are no left instances.  Note that the term "score" is *still*
used, but it refers to the pair (points, penalty).  When I say there  are no
left instance, I mean that the word "score" should not be used to refer to the
number of points on its own.